Refaktorering v3 -> v2 og V2 -> V1
