+++ src/main/java/no/nav/pto/veilarbportefolje/database/BrukerDataService.java
@@ -6,8 +6,8 @@ import no.nav.common.types.identer.AktorId;
 import no.nav.pto.veilarbportefolje.aktiviteter.*;
 import no.nav.pto.veilarbportefolje.arenapakafka.aktiviteter.GruppeAktivitetRepository;
 import no.nav.pto.veilarbportefolje.arenapakafka.aktiviteter.GruppeAktivitetRepositoryV2;
+import no.nav.pto.veilarbportefolje.arenapakafka.aktiviteter.TiltakRepositoryV1;
 import no.nav.pto.veilarbportefolje.arenapakafka.aktiviteter.TiltakRepositoryV2;
-import no.nav.pto.veilarbportefolje.arenapakafka.aktiviteter.TiltakRepositoryV3;
 import no.nav.pto.veilarbportefolje.arenapakafka.arenaDTO.GruppeAktivitetSchedueldDTO;
 import no.nav.pto.veilarbportefolje.arenapakafka.ytelser.YtelseDAO;
 import no.nav.pto.veilarbportefolje.arenapakafka.ytelser.YtelsesStatusRepositoryV2;
@@ -31,8 +31,8 @@ import static no.nav.pto.veilarbportefolje.config.FeatureToggle.brukIkkeAvtalteA
 @RequiredArgsConstructor
 public class BrukerDataService {
     private final AktivitetDAO aktivitetDAO;
+    private final TiltakRepositoryV1 tiltakRepositoryV1;
     private final TiltakRepositoryV2 tiltakRepositoryV2;
-    private final TiltakRepositoryV3 tiltakRepositoryV3;
     private final GruppeAktivitetRepository gruppeAktivitetRepository;
     private final GruppeAktivitetRepositoryV2 gruppeAktivitetRepositoryV2;
     private final BrukerDataRepository brukerDataRepository;
@@ -181,7 +181,7 @@ public class BrukerDataService {
 
     // TODO: vurder aa merge de to metodene under
     private List<Timestamp> hentAlleStartdatoer(AktorId aktorId, PersonId personId) {
-        List<Timestamp> startDatoer = tiltakRepositoryV2.hentStartDatoer(personId).stream()
+        List<Timestamp> startDatoer = tiltakRepositoryV1.hentStartDatoer(personId).stream()
                 .filter(Objects::nonNull).collect(toList());
         List<Timestamp> aktiviteter = aktivitetDAO.getAktiviteterForAktoerid(aktorId, brukIkkeAvtalteAktiviteter(unleashService)).getAktiviteter().stream()
                 .filter(AktivitetUtils::harIkkeStatusFullfort)
@@ -199,7 +199,7 @@ public class BrukerDataService {
 
 
     private List<Timestamp> hentAlleSluttdatoer(AktorId aktorId, PersonId personId) {
-        List<Timestamp> sluttdatoer = tiltakRepositoryV2.hentSluttdatoer(personId).stream()
+        List<Timestamp> sluttdatoer = tiltakRepositoryV1.hentSluttdatoer(personId).stream()
                 .filter(Objects::nonNull).collect(toList());
         List<Timestamp> aktiviteter = aktivitetDAO.getAktiviteterForAktoerid(aktorId, brukIkkeAvtalteAktiviteter(unleashService)).getAktiviteter().stream()
                 .filter(AktivitetUtils::harIkkeStatusFullfort)
@@ -225,7 +225,7 @@ public class BrukerDataService {
         List<Timestamp> gruppeAktiviteter = gruppeAktivitetRepositoryV2.hentAktiveAktivteter(aktorId).stream()
                 .map(GruppeAktivitetSchedueldDTO::getAktivitetperiodeFra)
                 .filter(Objects::nonNull).collect(toList());
-        List<Timestamp> tiltak = tiltakRepositoryV3.hentStartDatoer(aktorId).stream()
+        List<Timestamp> tiltak = tiltakRepositoryV2.hentStartDatoer(aktorId).stream()
                 .filter(Objects::nonNull).collect(toList());
         startDatoer.addAll(tiltak);
         startDatoer.addAll(gruppeAktiviteter);
@@ -243,7 +243,7 @@ public class BrukerDataService {
         List<Timestamp> gruppeAktiviteter = gruppeAktivitetRepositoryV2.hentAktiveAktivteter(aktorId).stream()
                 .map(GruppeAktivitetSchedueldDTO::getAktivitetperiodeTil)
                 .filter(Objects::nonNull).collect(toList());
-        List<Timestamp> tiltak = tiltakRepositoryV3.hentSluttdatoer(aktorId).stream()
+        List<Timestamp> tiltak = tiltakRepositoryV2.hentSluttdatoer(aktorId).stream()
                 .filter(Objects::nonNull).collect(toList());
         sluttdatoer.addAll(tiltak);
         sluttdatoer.addAll(gruppeAktiviteter);
