+++ src/test/java/no/nav/pto/veilarbportefolje/arenapakafka/TiltakPostgresTest.java
@@ -11,8 +11,8 @@ import no.nav.pto.veilarbportefolje.aktiviteter.AktiviteterRepositoryV2;
 import no.nav.pto.veilarbportefolje.aktiviteter.AktoerAktiviteter;
 import no.nav.pto.veilarbportefolje.arenapakafka.aktiviteter.GruppeAktivitetRepository;
 import no.nav.pto.veilarbportefolje.arenapakafka.aktiviteter.GruppeAktivitetRepositoryV2;
+import no.nav.pto.veilarbportefolje.arenapakafka.aktiviteter.TiltakRepositoryV1;
 import no.nav.pto.veilarbportefolje.arenapakafka.aktiviteter.TiltakRepositoryV2;
-import no.nav.pto.veilarbportefolje.arenapakafka.aktiviteter.TiltakRepositoryV3;
 import no.nav.pto.veilarbportefolje.arenapakafka.arenaDTO.TiltakInnhold;
 import no.nav.pto.veilarbportefolje.arenapakafka.ytelser.YtelsesStatusRepositoryV2;
 import no.nav.pto.veilarbportefolje.config.ApplicationConfigTest;
@@ -47,7 +47,7 @@ import static org.mockito.Mockito.mock;
 public class TiltakPostgresTest {
     private final JdbcTemplate db;
     private final OppfolginsbrukerRepositoryV2 oppfolginsbrukerRepositoryV2;
-    private final TiltakRepositoryV3 tiltakRepositoryV3;
+    private final TiltakRepositoryV2 tiltakRepositoryV2;
     private final AktivitetStatusRepositoryV2 aktivitetStatusRepositoryV2;
     private final BrukerDataService brukerDataService;
 
@@ -59,14 +59,14 @@ public class TiltakPostgresTest {
         db = SingletonPostgresContainer.init().createJdbcTemplate();
         oppfolginsbrukerRepositoryV2 = new OppfolginsbrukerRepositoryV2(db);
         aktivitetStatusRepositoryV2 = new AktivitetStatusRepositoryV2(db);
-        tiltakRepositoryV3 = new TiltakRepositoryV3(db, aktivitetStatusRepositoryV2);
+        tiltakRepositoryV2 = new TiltakRepositoryV2(db, aktivitetStatusRepositoryV2);
         YtelsesStatusRepositoryV2 ytelsesStatusRepositoryV2 = new YtelsesStatusRepositoryV2(db);
 
         GruppeAktivitetRepositoryV2 gruppeAktivitetRepositoryV2 = mock(GruppeAktivitetRepositoryV2.class);
         AktiviteterRepositoryV2 aktiviteterRepositoryV2 = mock(AktiviteterRepositoryV2.class);
         Mockito.when(gruppeAktivitetRepositoryV2.hentAktiveAktivteter(any())).thenReturn(new ArrayList<>());
         Mockito.when(aktiviteterRepositoryV2.getAktiviteterForAktoerid(any(), anyBoolean())).thenReturn(new AktoerAktiviteter("1").setAktiviteter(new ArrayList<>()));
-        brukerDataService = new BrukerDataService(mock(AktivitetDAO.class), mock(TiltakRepositoryV2.class), tiltakRepositoryV3, mock(GruppeAktivitetRepository.class), gruppeAktivitetRepositoryV2, mock(BrukerDataRepository.class), aktiviteterRepositoryV2, aktivitetStatusRepositoryV2, ytelsesStatusRepositoryV2, mock(UnleashService.class));
+        brukerDataService = new BrukerDataService(mock(AktivitetDAO.class), mock(TiltakRepositoryV1.class), tiltakRepositoryV2, mock(GruppeAktivitetRepository.class), gruppeAktivitetRepositoryV2, mock(BrukerDataRepository.class), aktiviteterRepositoryV2, aktivitetStatusRepositoryV2, ytelsesStatusRepositoryV2, mock(UnleashService.class));
     }
 
     @BeforeEach
@@ -91,16 +91,16 @@ public class TiltakPostgresTest {
                 .setEndretDato(new ArenaDato("2021-01-01"))
                 .setAktivitetperiodeTil(new ArenaDato("1990-01-01"))
                 .setAktivitetid("TA-123456789");
-        tiltakRepositoryV3.upsert(innhold, aktorId);
+        tiltakRepositoryV2.upsert(innhold, aktorId);
 
-        tiltakRepositoryV3.utledOgLagreTiltakInformasjon(aktorId);
+        tiltakRepositoryV2.utledOgLagreTiltakInformasjon(aktorId);
         brukerDataService.oppdaterAktivitetBrukerDataPostgres(aktorId);
 
-        List<String> tiltak = tiltakRepositoryV3.hentBrukertiltak(aktorId);
+        List<String> tiltak = tiltakRepositoryV2.hentBrukertiltak(aktorId);
         Optional<AktivitetStatus> aktivitetStatus = aktivitetStatusRepositoryV2.hentAktivitetTypeStatus(aktorId.get(), AktivitetTyper.tiltak.name());
         Optional<Timestamp> utloptAktivitet = aktivitetStatusRepositoryV2.hentAktivitetStatusUtlopt(aktorId.get());
 
-        Optional<String> kodeVerkNavn = tiltakRepositoryV3.hentVerdiITiltakskodeVerk(tiltaksType);
+        Optional<String> kodeVerkNavn = tiltakRepositoryV2.hentVerdiITiltakskodeVerk(tiltaksType);
 
         assertThat(tiltak.size()).isEqualTo(1);
         assertThat(tiltak.get(0)).isEqualTo(tiltaksType);
@@ -138,10 +138,10 @@ public class TiltakPostgresTest {
                 .setAktivitetperiodeTil(new ArenaDato(igarTid.toString().substring(0, 10)))
                 .setAktivitetid("TA-321");
 
-        tiltakRepositoryV3.upsert(idag, aktorId);
-        tiltakRepositoryV3.upsert(igar, aktorId);
+        tiltakRepositoryV2.upsert(idag, aktorId);
+        tiltakRepositoryV2.upsert(igar, aktorId);
 
-        tiltakRepositoryV3.utledOgLagreTiltakInformasjon(aktorId);
+        tiltakRepositoryV2.utledOgLagreTiltakInformasjon(aktorId);
         brukerDataService.oppdaterAktivitetBrukerDataPostgres(aktorId);
 
         Optional<AktivitetStatus> aktivitetStatus = aktivitetStatusRepositoryV2.hentAktivitetTypeStatus(aktorId.get(), AktivitetTyper.tiltak.name());
@@ -164,17 +164,17 @@ public class TiltakPostgresTest {
                 .setTiltakstype(tiltaksType)
                 .setDeltakerStatus("GJENN")
                 .setAktivitetid(id);
-        tiltakRepositoryV3.upsert(innhold, aktorId);
-        tiltakRepositoryV3.delete(id);
+        tiltakRepositoryV2.upsert(innhold, aktorId);
+        tiltakRepositoryV2.delete(id);
 
-        tiltakRepositoryV3.utledOgLagreTiltakInformasjon(aktorId);
+        tiltakRepositoryV2.utledOgLagreTiltakInformasjon(aktorId);
         brukerDataService.oppdaterAktivitetBrukerDataPostgres(aktorId);
 
-        List<String> tiltak = tiltakRepositoryV3.hentBrukertiltak(aktorId);
+        List<String> tiltak = tiltakRepositoryV2.hentBrukertiltak(aktorId);
         Optional<AktivitetStatus> aktivitetStatus = aktivitetStatusRepositoryV2.hentAktivitetTypeStatus(aktorId.get(), AktivitetTyper.tiltak.name());
         Optional<Timestamp> utloptAktivitet = aktivitetStatusRepositoryV2.hentAktivitetStatusUtlopt(aktorId.get());
 
-        Optional<String> kodeVerkNavn = tiltakRepositoryV3.hentVerdiITiltakskodeVerk(tiltaksType);
+        Optional<String> kodeVerkNavn = tiltakRepositoryV2.hentVerdiITiltakskodeVerk(tiltaksType);
 
         assertThat(tiltak.size()).isEqualTo(0);
         assertThat(kodeVerkNavn.isPresent()).isTrue();
@@ -209,10 +209,10 @@ public class TiltakPostgresTest {
                 .setDeltakerStatus("GJENN")
                 .setAktivitetid("T-321");
 
-        tiltakRepositoryV3.upsert(tiltak1, aktorId);
-        tiltakRepositoryV3.upsert(tiltak2, aktorId);
+        tiltakRepositoryV2.upsert(tiltak1, aktorId);
+        tiltakRepositoryV2.upsert(tiltak2, aktorId);
 
-        EnhetTiltak enhetTiltak = tiltakRepositoryV3.hentTiltakPaEnhet(EnhetId.of(navKontor));
+        EnhetTiltak enhetTiltak = tiltakRepositoryV2.hentTiltakPaEnhet(EnhetId.of(navKontor));
         assertThat(enhetTiltak.getTiltak().size()).isEqualTo(2);
         assertThat(enhetTiltak.getTiltak().get(tiltaksType1)).isEqualTo(tiltaksNavn1);
         assertThat(enhetTiltak.getTiltak().get(tiltaksType2)).isEqualTo(tiltaksNavn2);
