+++ src/test/java/no/nav/pto/veilarbportefolje/config/ApplicationConfigTest.java
@@ -17,8 +17,8 @@ import no.nav.pto.veilarbportefolje.arbeidsliste.ArbeidslisteService;
 import no.nav.pto.veilarbportefolje.arenapakafka.aktiviteter.ArenaHendelseRepository;
 import no.nav.pto.veilarbportefolje.arenapakafka.aktiviteter.GruppeAktivitetRepository;
 import no.nav.pto.veilarbportefolje.arenapakafka.aktiviteter.GruppeAktivitetRepositoryV2;
+import no.nav.pto.veilarbportefolje.arenapakafka.aktiviteter.TiltakRepositoryV1;
 import no.nav.pto.veilarbportefolje.arenapakafka.aktiviteter.TiltakRepositoryV2;
-import no.nav.pto.veilarbportefolje.arenapakafka.aktiviteter.TiltakRepositoryV3;
 import no.nav.pto.veilarbportefolje.arenapakafka.aktiviteter.UtdanningsAktivitetService;
 import no.nav.pto.veilarbportefolje.arenapakafka.ytelser.YtelsesRepository;
 import no.nav.pto.veilarbportefolje.arenapakafka.ytelser.YtelsesRepositoryV2;
@@ -135,8 +135,8 @@ import static org.mockito.Mockito.when;
         ArenaHendelseRepository.class,
         GruppeAktivitetRepository.class,
         GruppeAktivitetRepositoryV2.class,
+        TiltakRepositoryV1.class,
         TiltakRepositoryV2.class,
-        TiltakRepositoryV3.class,
         BrukerDataService.class,
         BrukerDataRepository.class,
         YtelsesService.class,
